+++ js/components/accordion.js
@@ -31,7 +31,7 @@ Toolkit.Accordion = Toolkit.Component.extend({
         this.options = options = this.setOptions(options, element);
 
         // Find headers and cache the index of each header and set ARIA attributes
-        this.headers = element.find('.' + Toolkit.vendor + 'accordion-header').each(function(index) {
+        this.headers = element.find('[data-accordion-header]').each(function(index) {
             $(this)
                 .data('accordion-index', index)
                 .attr({
@@ -46,7 +46,7 @@ Toolkit.Accordion = Toolkit.Component.extend({
         });
 
         // Find sections and cache the height so we can use for sliding and set ARIA attributes
-        this.sections = element.find('.' + Toolkit.vendor + 'accordion-section').each(function(index) {
+        this.sections = element.find('[data-accordion-section]').each(function(index) {
             $(this)
                 .data('height', $(this).height())
                 .attr({
@@ -59,7 +59,7 @@ Toolkit.Accordion = Toolkit.Component.extend({
 
         // Set events
         this.events = {
-            '{mode} element .@accordion-header': 'onShow'
+            '{mode} element [data-accordion-header]': 'onShow'
         };
 
         // Initialize